[HUDI-717] Fixed usage of HiveDriver for DDL statements.
When using the HiveDriver mode in HudiHiveClient, Hive 2.x DDL operations like ALTER PARTITION may fail. This is because Hive 2.x doesn't like `db`.`table_name` for operations.

In this fix, we set the name of the database in the SessionState create for the Driver.

Added a unit test for the fix as well as additional unit tests for HoodieHiveClient.