[global-hive-sync-tool] Add a global hive sync tool to sync hoodi table across clusters. Add a way to rollback the replicated time stamp if we fail to sync or if we partly sync
Summary:
Note that right now we are targetting to rollback only the replicated timestamp. We are not
rollingback any new schema or partition changes that may have partly synced into hive. There
wont be any data visible for new partitions but it is possible that schema changes are not
backward compatible.

In a later change we will introduce a way to run hive sync tool to restore the partitions
to earlier state.

Test Plan: added unit tests