[HUDI-1598] Write as minor batches during one checkpoint interval for the new writer
This is the #step 3 of RFC-24:
https://cwiki.apache.org/confluence/display/HUDI/RFC+-+24%3A+Hoodie+Flink+Writer+Proposal

Buffering data during one checkpoint when flush the buffer out all at a
time is not resource friendly for streaming write. The more proper way
it to cut the batches based on their real memory data buffer size (say,
128Mb), the writer always flushes the buffer out when its size reaches
the configured threshold.

Thus, after this change, one instant may span one (if every checkpoint
succeeds) or more (if there are checkpoint failures) checkpoints. The
instant only commits when there is a successful checkpoint.

Changes:
- Modify the BaseFlinkCommitActionExecutor to keep the write handles
  explicitly, the write task can then decides when and how to roll over
  the underneath file handles
- The StreamWriteOperatorCoordinator now commits the old instant and
  creates new one only if the checkpoint succeeds, thus, we can skip
  some failed checkpoints actually.