[HUDI-1624] The state based index should bootstrap from existing base files
The index should bootstrap from existing base files if the files exist, in the
design, we load all the keys for one partition if we found that the key
does not exist in the index for `processElement`, if there are many
records for this partition, the processing may block and trigger back
pressure. When all the records are loaded, we only need to check the
state each time a record is tagged.