[UBER] Added the drogon conf file for delta streamer
Summary:
We need to add the drogon JSON file to invoke deltastremer in the Uber ecosystem, this is the initial version of the drogon file with few default values.

The user should pass the correct variable values via Piper/uWorc before invoking the drogon launch command.

Drogon build will create the build artifacts under the .m2 folder hence added that folder to the .gitignore file.

This is the current master version hence used it as the default value: 0.6.1-SNAPSHOT

Test Plan:
Tested extensively with drogon build && drogon deploy && drogon launch commands, here is one of the executions:

https://shs-phx2.uberinternal.com/history/application_1615489146956_933030/1/jobs/

Reviewers: satishkotha, modi, O955 Project Hoodie Project Reviewer: Add blocking reviewers, PHID-PROJ-pxfpotkfgkanblb3detq

Reviewed By: satishkotha, O955 Project Hoodie Project Reviewer: Add blocking reviewers, PHID-PROJ-pxfpotkfgkanblb3detq

JIRA Issues: HUDI-592

Differential Revision: https://code.uberinternal.com/D5802949