[HUDI-251] Adds JDBC source support for DeltaStreamer
As discussed in RFC-14, this change implements the first phase of JDBC incremental puller.
It consists following changes:

- JdbcSource: This class extends RowSource and implements
  fetchNextBatch(Option<String> lastCkptStr, long sourceLimit)

- SqlQueryBuilder: A simple utility class to build sql queries fluently.

- Implements two modes of fetching: full and incremental.
  Full is a complete scan of RDBMS table.
  Incremental is delta since last checkpoint.
  Incremental mode falls back to full fetch in case of any exception.

- Caches the dataset after scan. Storage level is configurable.

[HUDI-251] Adds JDBC source support for DeltaStreamer

This is the second revision of the PR. It has following changes:

- Make the decision to fallback to full fetch mode
  configurable when incremental fetch fails.

- JdbcTestUtils: Helper class to help with sql execution
  commonly used in JdbcSource setup.

- Added more tests for JdbcSource and a long-running
  incremental fetch test in TestHoodieDeltaStreamer.

- Addressed a few other minor comments.

This is the third revision, which addresses some of the minor comments.

This is the fourth revision which makes the checkpoint return the
last known checkpoint in case of exception. It also renames some
of the config keys as suggested in review comment.

Make checkpoint throw exception and stop sync if it fails.