[UBER] Remove cache call from pre-commit validation phase and cache the sql results to avoid shuffle
Summary:
Caching before and after state of the entire partition is causing OOM issues. Also since we do a join operation for before and after state it is triggering a shuffle operation both the states. In general, when precommit are configured max one of them is configured and the validation SQLs may contain few columns so this commit will basically run the SQLs first for both before and after state cache them and then run join operation on the results. This way the shuffle needs to deal with very less data and it is much faster.

Test Plan:
The changes are tested on a temporary table. Spark job details are as follows.

Spark job:
https://shs-phx2.uberinternal.com/history/application_1641417840429_1541644

Reviewers: meenalb, bkrishen, O955 Project Hoodie Project Reviewer: Add blocking reviewers, PHID-PROJ-pxfpotkfgkanblb3detq!

Reviewed By: meenalb, O955 Project Hoodie Project Reviewer: Add blocking reviewers

JIRA Issues: HUDI-1691

Differential Revision: https://code.uberinternal.com/D7033921