[HUDI-5403] Turn off metadata-table-based file listing in BaseHoodieTableFileIndex (#7488)
Currently, on the reader or query engine side, the direct file listing on the file system is used by default, as indicated by HoodieMetadataConfig.DEFAULT_METADATA_ENABLE_FOR_READERS (=false). Without providing explicit config of hoodie.metadata.enable, the metadata-table-based file listing is disabled. However, the BaseHoodieTableFileIndex, the common File Index implementation, used by Trino Hive connector, does not respect this default behavior. This leads to performance regression of query latency in Trino Hive connector, due to way of how the connector is integrated with the Input Format and the File Index with metadata table enabled.

This PR fixes the BaseHoodieTableFileIndex to respect the expected behavior defined by HoodieMetadataConfig.DEFAULT_METADATA_ENABLE_FOR_READERS, i.e., metadata-table-based file listing is disabled by default. The metadata-table-based file listing is only enabled when hoodie.metadata.enable is set to true and the files partition of the metadata table is ready for read based on the Hudi table config.

Impact
This mitigates the performance regression of query latency in Trino Hive connector and fixes the read-side behavior of the file listing.

Tested the PR that by default, the HoodieParquetInputFormat does not read metadata table for file listing anymore.

Co-authored-by: Sagar Sumit <sagarsumit09@gmail.com>
Co-authored-by: Alexey Kudinkin <alexey@infinilake.com>