[HUDI-5511] Do not clean the CkpMetadata dir when restart the job
At the beginning, we bootstrap the ckp metadata by cleaning all the messages.
This brings in some corner case like 'the write task can not fetch the pending instant correctly when restartting the job',
if a checkpoint succeed and the job crashes suddenly, the instant hasn't had time to commit, then the data loss happens,
because the last pending instant would be rolled back, while the Flink engin thinks the checkpint/instant is successful.

Q: Why we clean the messages here ?
A: To prevent inconsistencies between timeline and the messages.

Q: Why we decide to keep the messages ?
A: There are two cases for the inconsistency:

1. the timeline instant is complete but the ckp message is inflight (for committing instant);
2. the timeline instant is pending while the ckp message does not start (for starting a new instant);

For case1, there is no need to re-commit the instant, so it's okey the write task does not get any pending instant when recovering,
for case2, the instant is basically pending, it would be rolled back which is in line with expectations.
so keeping the ckp messages as is can actually maintain correctness.