[HUDI-5349] Clean up partially failed restore (#7605)
If a table was attempted w/ "restore" operation and if it failed mid-way, restore could still be lying around. when re-attempted, a new instant time will be allotted and re-attempted from scratch. but this may thwart compaction progression in MDT. so we need to ensure for a given savepoint, we always re-use restore instant if any. 

Co-authored-by: Jonathan Vexler <=>