[HUDI-5589] Fix Hudi config inference (#7713)
For a config defined by `ConfigProperty` with no default value set, even if the inference function is set with `.withInferFunction()` in `HoodieConfig`, the inference function is not invoked in `HoodieConfig` to generate the config value. This commit fixes the issue so that such a config can still get the inferred value without a default value set.