[HUDI-1575] Early Conflict Detection For Multi-writer (#6133)
Before this PR, Hudi implements OCC (Optimistic Concurrency Control) to detect the write conflict at the pre-commit time to ensure data consistency, integrity, and correctness between multiple writers. OCC detects the conflict at Hudi's file group level, i.e., two concurrent writers updating the same file group are detected as a conflict.  Currently, conflict detection is performed before committing metadata and after the data writing is completed. If any conflict is detected, it leads to a waste of cluster resources because computing and writing are finished already.

To solve this problem, this change implements an early conflict detection mechanism to detect the conflict during the data writing phase and abort the writing early if a conflict is detected, using Hudi's marker mechanism. Before writing each data file, the writer creates a corresponding marker to mark that the file is created, so that later on, the writer can use the markers to automatically clean up uncommitted data files in the failure and rollback scenarios. We leverage the markers to identify the conflict at the file group level during writing data. There are subtle differences in the early conflict detection workflow among different types of markers.  For direct markers, the writer lists necessary marker files directly and checks the conflict before creating markers and starting to write the corresponding data file. For the timeline-server-based markers, the writer gets the result of marker conflict detection by contacting the timeline server. The conflict detection is asynchronously and periodically executed at the timeline server so that the write conflicts can be detected as early as possible.  Both writers may still write the data files of the same file slice until the conflict is detected in the next round of detection.

Note that, the implemented early conflict detection operates within OCC. Any conflict detection outside the scope of OCC is not handled.  For example, the current OCC for multiple writers cannot detect the conflict if two concurrent writers perform INSERT operations for the same set of record keys, because the writers write to different file groups. This set of changes does not intend to address this problem.

Co-authored-by: yuezhang <yuezhang@freewheel.tv>
Co-authored-by: Y Ethan Guo <ethan.guoyihua@gmail.com>