[HUDI-5401] Ensure user-provided hive metastore uri is set in HiveConf if not already set (#7543)
If user has set hoodie.datasource.hive_sync.metastore.uris to a remote metastore but hive.metastore.uris in hadoop conf is not set, then sync will succeed to embedded metastore and the tables won't be visible in remote metastore. This makes the user think that sync was successful, yet tables are not found. This PR fixes that behavior.