[HUDI-5485] Add File System View API for batch listing and improve savepoint performance with metadata table (#7690)
The savepoint action needs to list all the latest base files before or on the target instant time. When the metadata table is enabled, for each partition, the metadata table is scanned from the Spark executor, which leads to a lot of S3 requests. This significantly degrades the savepoint performance when the number of table partitions is huge. Such repeated scanning of the metadata table is unnecessary and can be avoided by scanning the metadata table once.

This commit fixes the issue by adding a new file system view API that batches the file listing of all partitions so that only one metadata table lookup call (`tableMetadata.getAllFilesInPartitions`) is invoked on the driver side for savepoint action.