[HUDI-5704] De-coupling column drop flag and schema validation flag (#7857)
De-coupling column drop flag and schema validation flag. Looks like we had tight coupling before.
for eg,
if table schema is col1, col2, col3
and new incoming schema is col1, col2
col drop config is set to false (which means col drop should not be supported), and schema validation is set to false, commit will succeed. Expectation is, commit should fail for this new batch.

Fixing the use-case in this patch, where we are de-coupling them. col drop flag will be honored irrespective of whether schema validation is enabled or not.
---------

Co-authored-by: Sagar Sumit <sagarsumit09@gmail.com>