[HUDI-5795] Ensure we only have a single push gw client and hence a single connection pool
Currently we maintain a single Metrics instance per table (and a separate one for metadata table). That means we create a separate instance for PushGateway, the client for push gateway. This implies that we have a connection pool per hudi table, that is leading to a lot of open/waiting/blocked TCP connections on the push gw, possibly causing it to crash.

In this PR, we ensure that we use a single PushGateway instance per hostname and port, ensuring we only maintain a single connection pool per PushGw Server Instance