[HUDI-4442] [HUDI-5001] Sanitize JsonConversion and RowSource (#8010)
Updates the MercifulJsonConverter to have more mercy by:

Replaces field names with illegal avro name chars with __
Leverages aliases in the avro schema if present
This issue is within the scope of row sources. The actual issue is that if the names of the columns in the row sources contain invalid avro characters ref [here|https://avro.apache.org/docs/1.10.2/spec.html#names] then using configuration set we can sanitize the column names both in the schema and actual data and the data ingestion to hudi isn't failed. The schema provider is scoped out to filebasedschemaregistry as other schema registries might not allow to register invalid schema in the first place.

---------

Co-authored-by: Jonathan Vexler <=>