[HUDI-5863] Fix HoodieMetadataFileSystemView serving stale view at the timeline server (#8079)
We observe that for MOR table, occasionally (<10% for large tables with frequent updates and compactions), the deltacommit after the compaction commit may add a new log file to the old file slice, not the latest file slice, in the corresponding file group.  This happens when both the metadata table and timeline server are enabled.  If either is disabled, the problem does not show up. This means that any data written to the log file is lost, i.e., data loss issue.

Deeper analysis of the code surfaces that the file system view at the timeline server may serve the stale view, causing the issue.  This is because the sync of `HoodieMetadataFileSystemView` is not atomic when the metadata table is enabled.

This commit makes the following fixes:
- Makes sure all logic in `sync()` and `reset()` are guarded by the write lock.  This is done by implementing the logic with the write lock in each file system view, instead of overriding indect method (e.g., `runSync()`).
- Simplifies the implementation of `syncIfLocalViewBehind()` to be more readable.