[HUDI-5891] Fix clustering on bootstrapped tables (#8206)
- Fixes clustering on bootstrapped tables. There are two issues:

Clustering strategy added bootstrap as well as source paths to the glob paths. This was causing an error, as mentioned in HUDI-5891, while building file index. There is no need to add source path as that is handled in HoodieBootstrapRDD.
Sometimes the path will have scheme and sometimes it will be without scheme, so getting the relative partition path based was buggy. Corrected that in HoodieBootstrapRelation.