[HUDI-5989] Fix date conversion issue when performing partition pruning on Spark (#8298)
When lazy fetching partition path & file slice for HoodieFileIndex is used, date cannot be converted to the correct string representation.

This is the case as Spark store dates as an integer value representing the number of days that has past since 1970-01-01.

When rebuilding the partition path, this FS path could be rebuilt wrongly causing a partition to be empty, and hence, the query result to be empty/incorrect.

Co-authored-by: Rex An <bonean131@gmail.com>