[HUDI-5723] Automate and standardize enum configs (#7881)
This commit creates standard annotation for enums that is used to set config defaults, restrict input choices, and automate documentation.  This should make code more readable by adding comments to many enums that previously had none.  The commit additionally rewords many of the annotations to be readable so our documentation is more clear.

---------

Co-authored-by: Jonathan Vexler <=>
Co-authored-by: Y Ethan Guo <ethan.guoyihua@gmail.com>
Co-authored-by: Lokesh Jain <ljain@apache.org>