[HUDI-6186] Fix lock identity in InProcessLockProvider (#8658)
This commit fixes a bug introduced by #6847. #6847 extends the InProcessLockProvider to support multiple tables in the same process, by having an in-memory static final map storing the mapping of the table base path to the read-write reentrant lock, so that the writer uses the corresponding lock based on the base path. When closing the lock provider, close() removes the lock entry. Since close() is called when closing the write client, the lock is removed and subsequent concurrent writers will get a different lock instance on the same table, causing the locking mechanism on the same table to be useless.  The fix gets rid of the lock removal operation in the `close()` call since it has to be kept for concurrent writers.  A new test `TestInProcessLockProvider#testLockIdentity` based on the above scenario is added to guard the behavior.