[HUDI-6199] Fix deletes with custom payload implementation (#8690)
There was a bug that the delete records are assumed to be marked by "_hoodie_is_deleted"; however, custom CDC payloads use "op" field to mark deletes.  In such a case, AWS DMS payload and Debezium payload failed with deletes.  This commit fixes the issue by adding a new API isDeleteRecord(GenericRecord genericRecord) in BaseAvroPayload to allow the payload to implement custom logic to indicate if a record is a delete record.

Co-authored-by: Raymond Xu <2701446+xushiyan@users.noreply.github.com>