[HUDI-6245] Automatically downgrade table version of metadata table (#8771)
This commit adds the logic to automatically downgrade or upgrade the table version of the metadata table to improve the user experience. Before this commit, the user has to manually downgrade the metadata table from version 5 to 4 (e.g., from 0.13.0 release to 0.11.1 release), when the metadata table is enabled, along with the data table.