[HUDI-6391] Fix pre-commit validators based on Spark SQL (#8993)
This commit fixes the pre-commit validatorsbased on Spark SQL (SqlQueryInequalityPreCommitValidator and SqlQueryEqualityPreCommitValidator). Before this commit, for the before state in the validation, the getRecordsFromCommittedFiles returns an empty data frame without a schema, causing SQL with column name reference to fail. This commit fixes the issue by either using the table schema if present, or the schema from after state, if the table is empty.