[HUDI-6462] Add Hudi client init callback interface (#9108)
This PR adds the interface for Hudi client init callback to run custom logic at the time of initialization of a Hudi client:

@PublicAPIClass(maturity = ApiMaturityLevel.EVOLVING)
public interface HoodieClientInitCallback {
  /**
   * A callback method in which the user can implement custom logic.
   * This method is called when a {@link BaseHoodieClient} is initialized.
   *
   * @param hoodieClient {@link BaseHoodieClient} instance.
   */
  @PublicAPIMethod(maturity = ApiMaturityLevel.EVOLVING)
  void call(BaseHoodieClient hoodieClient);
}
At the time of instantiation of the write or table service client, a user may want to do additional processing, such as sending metrics, logsm notification, or adding more properties to the write config. The implementation of client init callback interface allows such logic to be plugged into Hudi.

A new config, hoodie.client.init.callback.classes, is added for plugging in the callback implementation. The class list is comma-separated.