[HUDI-6464] Spark SQL Merge Into for pkless tables (#9083)
Tables with a primary key now must join on all primary key columns. Additionally, they can join on the partition path columns as well, which is recommended if the table is not using a global index.

Tables without a primary key can join on any columns in the table. If multiple source table columns match a single target table column, precombine field will be used if set; otherwise, behavior is nondeterminate. To improve performance, the hudi meta cols are retained after the join, so that index lookup and keygeneration can be skipped.