[HUDI-6153] Changed the rollback mechanism for MDT to actual rollbacks rather than appending revert blocks. (#8837)
Prior to this patch, rollbacks in data table (DT) is yet another delta commit 
in metadata table (MDT), and we mark all files as invalid in that delta commit. 
But as we are adding more high-scale partitions like record indec, we can't 
afford to mark 100k entries as invalid. So, we wanted to trigger an actual 
rollback of the commit of interest in MDT for the delta commit of interest 
(similar to how we do rollback in DT).

---------

Co-authored-by: Sagar Sumit <sagarsumit09@gmail.com>
Co-authored-by: sivabalan <n.siva.b@gmail.com>