[HUDI-6609] Reverting multi writer checkpointing with HoodieStreamer (#9312)
Checkpointing with deltastreamer to support multiple deltastreamer writers might needs some format change. And we need to test backwards compatability. In the interest of time for 0.14.0, we are reverting this feature introduced by #8043. This multi-writer support for deltastreamer is not yet part of any official release yet. So, we should be good on this revert. Proper way to fix is covered in HUDI-6609.