[HUDI-6649] Fix column stat based data filtering for MOR (#9345)
Currently MOR snapshot relation does not use the column stats index for pruning 
the files in its queries. This PR aims to add support for pruning the file slices 
based on column stats in case of MOR.

The approach is similar to what was used for org.apache.hudi.HoodieFileIndex#listFiles. 
For every partition path, logic for filtering the file slices has been moved to function org.apache.hudi.HoodieFileIndex#filterFileSlices. This function is called by various 
relations for pruning the file slices.

---------

Co-authored-by: Sagar Sumit <sagarsumit09@gmail.com>